Methods (both styles) and Session-scope get queries now working for both V1 and V2. Almost all of the tests pass.
git-svn-id: https://svn.apache.org/repos/asf/qpid/branches/qmf-devel0.7a@929104 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/managementgen/qmfgen/templates/Class.cpp
@@ -333,6 +333,6 @@ void /*MGEN:Class.NameCap*/::doMethod (/*MGEN:Class.DoMapMethodArgs*/)
     std::string          text;
 
 /*MGEN:Class.MapMethodHandlers*/
-    outMap["_status_code"] = status;
+    outMap["_status_code"] = (uint32_t) status;
     outMap["_status_text"] = Manageable::StatusText(status, text);
 }
